backend/requirements.txt
Flask
psycopg2-binary
flask-swagger-ui
flask-cors
flask-jwt-extended
Pillow
requests
gtfs-realtime-bindings
pandas